fix: bug repeat login is wrong on flutter client script
fix: bug repeat login is wrong on flutter client script

+++ admin_client/lib/models/entity/runtime_statistics_res.dart
@@ -46,10 +46,10 @@
 import 'dart:convert';
 
 class RStatisticsRes{
-  int usedMemory;
-  int maxMemory;
-  int freeMemory;
-  int totalMemory;
+  double usedMemory;
+  double maxMemory;
+  double freeMemory;
+  double totalMemory;
   int threadRunning;
   int cpuUsage;
